refactor(chat): Phase D1 — 拆 chat_ui + slash 命令注册表
治本 god-file（cli.py）与 25 分支 slash if 链：

D1a 抽 chat_ui.py：把对话展示层 helper（_LiveSpinner / _StreamPrinter / _dwidth /
   _is_amazon_domain / _looks_like_code_task / 域词表）搬到独立模块，cli re-export
   保引用与测试兼容；cli.py 3276→约3150 行，零循环依赖。

D1b slash 命令注册表：把循环里 ~25 条 if line==... 改成 _SLASH_HANDLERS 字典 +
   单次 dispatch；handler 用 nonlocal 重绑定 messages/render_md/stream_live/instructions，
   自定义命令展开后仍正确贯穿到模型轮。加命令从此 = 一个函数 + 一个 dict 条目，
   不必再在巨型 if 链里找位置。if line 分支 29→10。

校验：ruff 干净；475 单测全过；pty 冒烟逐条验证 /help分组·/plan·/raw·/stream·
/auto-edit·/cost·/tools·/knowledge·未知命令·/clear 行为，且 /clear+/auto-edit on
后自然语言编辑任务零审批落盘——证明 nonlocal 重绑定在模型轮零漂移。
